refactor date operator parsing helpers
